requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
azure-ai-inference
azure-core
pandas
pillow
python-docx
openpyxl
python-pptx
PyPDF2
boto3
anthropic
langchain
langchain-community
sentence-transformers
unstructured
unstructured-inference
faiss-cpu
PyMuPDF
html2text
beautifulsoup4
mistralai
serpapi==0.1.5
lightgbm==4.4.0

